Change map_vis_z from float to double

+++ nav2_costmap_2d/include/nav2_costmap_2d/costmap_2d_publisher.hpp
@@ -73,7 +73,7 @@ public:
     std::string global_frame,
     std::string topic_name,
     bool always_send_full_costmap = false,
-    float map_vis_z = 0.0);
+    double map_vis_z = 0.0);
 
   /**
    * @brief  Destructor
@@ -167,7 +167,7 @@ private:
   double saved_origin_y_;
   bool active_;
   bool always_send_full_costmap_;
-  float map_vis_z_;
+  double map_vis_z_;
 
   // Publisher for translated costmap values as msg::OccupancyGrid used in visualization
   rclcpp_lifecycle::LifecyclePublisher<nav_msgs::msg::OccupancyGrid>::SharedPtr costmap_pub_;
